apps/embedding-worker/requirements.txt
openai>=1.40.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pgvector>=0.3.0
python-dotenv>=1.0.0
numpy>=1.26.0
